Some calls during `httpx.AsyncClient` init process are blocking and also very angry to memory. In the case of concurrent run of thousands coroutines that has own client instance, that instances start its own SSL init, which causes certs files IO etc. So, using property that stores once initialized httpx.AsyncClient instance is a better way. But it makes using AsyncClient's context manager senseless, so we need to explicitly close that client after using.
This is a first try to solve the problem. Maybe the AsyncFirebaseClient context manager is a better choice. It helps to implicitly pre-cache access token, that is another problem in concurrent execution, as at the moment of coro is scheduled, _get_access_token not yet called and will be called by each coroutine

So, performance related workarounds includes:
 - loading all conf and creds files at app startup or during instance construction
 - make SSL context in the same way and pass it to client
 - init single httpx client outside coroutines (e.g. in the class's constructor)
 - pre-populate single instance of httpx client with access token before scheduling a task that uses client